Refactor: move each component from post into own file for better overview
