+++ app/components/ui/reply-thread.tsx
@@ -1,4 +1,4 @@
-import { PostDetails } from '#app/components/ui/post.tsx'
+import { PostDetails } from '#app/components/ui/post-details.tsx'
 import { CONVINCINGNESS_THRESHOLD } from '#app/constants.ts'
 import { type ThreadPost } from '#app/conversations.ts'
 import { Direction, type VoteState } from '#app/vote.ts'
