+++ app/routes/index.tsx
@@ -4,8 +4,9 @@ import moment from 'moment'
 import { useState } from 'react'
 import { Markdown } from '#app/components/markdown.tsx'
 import { Button } from '#app/components/ui/button.tsx'
+import { CommentIcon } from '#app/components/ui/comment-icon.tsx'
+import { PostContent } from '#app/components/ui/post-content.tsx'
 import { PostForm } from '#app/components/ui/post-form.tsx'
-import { PostContent, CommentIcon } from '#app/components/ui/post.tsx'
 import { db } from '#app/db.ts'
 import * as rankingTs from '#app/ranking.ts'
 import { getUserId } from '#app/utils/auth.server.ts'
