+++ app/routes/old+/post.$postId.tsx
@@ -6,7 +6,7 @@ import { GeneralErrorBoundary } from '#app/components/error-boundary.tsx'
 import { DeletedPost } from '#app/components/ui/deleted-post.tsx'
 import { DirectReplies } from '#app/components/ui/direct-replies.tsx'
 import { ParentThread } from '#app/components/ui/parent-thread.tsx'
-import { PostDetails } from '#app/components/ui/post.tsx'
+import { PostDetails } from '#app/components/ui/post-details.tsx'
 import { ReplyThread } from '#app/components/ui/reply-thread.tsx'
 import { getCriticalThread, type ThreadPost } from '#app/conversations.ts'
 import { db } from '#app/db.ts'
